requirements.txt
asgiref==3.5.0
autopep8==1.6.0
coverage==6.3.2
Django==4.0.3
django-cors-headers==3.11.0
django-countries==7.3.2
django-js-asset==2.0.0
django-mptt==0.13.4
django-taggit==2.1.0
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
Faker==13.3.1
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
tzdata==2021.5
